Something about stacking

+++ util.py
@@ -5,9 +5,10 @@ from PIL import Image
 
 def load_image(filename):
 	image = Image.open(filename).convert('L') # to grayscale
-	return np.asarray(image, dtype=np.int8)
+	return np.asarray(image, dtype=np.int16)
 
 
 def save_image(array, filename):
+	array = array.astype(np.int8)
 	image = Image.fromarray(array, mode='L')
 	image.save(filename)
